requirements.txt
ZConfig==2.9.2
ZODB3==3.10.5
transaction==1.3.0
zc.lockfile==1.0.0
zdaemon==3.0.1
zope.event==4.0.0
zope.interface==4.0.1
Flask-Script==0.3.3
Flask==0.8
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
path.py==2.2.2
CherryPy==3.2.2
python-ldap==2.4.10
blinker==1.2
